Repo: fix linters and organize the code base
